Refactor: move downcast logic to dataframe_util, add version gate
- Move _downcast_large_type and _downcast_large_arrow_types to
  dataframe_util.py to avoid duplicating convert_pandas_df_to_arrow_bytes
  logic in component_arrow.py
- Add downcast_large_types parameter to convert_pandas_df_to_arrow_bytes
- Gate downcasting behind pandas >= 3.0.0 version check in
  component_arrow to avoid touching schemas on older pandas versions
- Simplify component_arrow._convert_df_to_component_arrow_bytes to
  delegate to the shared helper
- Add unit tests for the new parameter in dataframe_util_test.py
